refactor(TodaysVideo): move state video to Index
